Refactor bbq format to be contained in a package
